Defining Schemas
Schemas designed for generality to allow adoption in different
countries/localities.

A Country is introduced to allow different countries with different
transport systems to be able to customize the system and use it.

A Transport will allow different types of transport systems such as
Road and Railway to be available in the Service.